ml-service/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1
httpx==0.26.0

# Phase 2 NLP (installed now so the image is ready)
spacy==3.7.4
nltk==3.8.1
scikit-learn==1.4.0
pdfminer.six==20231228
python-docx==1.1.0
pandas==2.2.0
numpy==1.26.3
language-tool-python==2.7.1
